+++ packages/eyes-selenium/lib/fluent/AccessibilityRegionByElement.js
@@ -5,6 +5,7 @@ const {
   AccessibilityMatchSettings,
   Location,
   CoordinatesType,
+  EyesJsBrowserUtils,
 } = require('@applitools/eyes-sdk-core')
 
 /**
@@ -43,6 +44,17 @@ class AccessibilityRegionByElement extends GetAccessibilityRegion {
       type: this._regionType,
     })
     return [accessibilityRegion]
+  }
+
+  async toPersistedRegions(driver) {
+    const xpath = await EyesJsBrowserUtils.getElementXpath(driver, this._element)
+    return [
+      {
+        type: 'xpath',
+        selector: xpath,
+        accessibilityType: this._regionType,
+      },
+    ]
   }
 }