+++ packages/eyes-selenium/lib/fluent/FloatingRegionBySelector.js
@@ -5,6 +5,7 @@ const {
   Location,
   CoordinatesType,
   FloatingMatchSettings,
+  locatorToPersistedRegions,
 } = require('@applitools/eyes-sdk-core')
 
 /**
@@ -60,6 +61,17 @@ class FloatingRegionBySelector extends GetFloatingRegion {
     }
 
     return values
+  }
+
+  async toPersistedRegions(driver) {
+    const regions = await locatorToPersistedRegions(this._selector, driver)
+    return regions.map(reg => ({
+      ...reg,
+      maxUpOffset: this._maxUpOffset,
+      maxDownOffset: this._maxDownOffset,
+      maxLeftOffset: this._maxLeftOffset,
+      maxRightOffset: this._maxRightOffset,
+    }))
   }
 }