+++ packages/eyes-selenium/lib/fluent/SelectorByLocator.js
@@ -1,6 +1,6 @@
 'use strict'
 
-const {GetSelector, GeneralUtils} = require('@applitools/eyes-sdk-core')
+const {GetSelector, GeneralUtils, locatorToPersistedRegions} = require('@applitools/eyes-sdk-core')
 
 const EYES_SELECTOR_TAG = 'data-eyes-selector'
 
@@ -39,6 +39,10 @@ class SelectorByLocator extends GetSelector {
     }
 
     return `[${EYES_SELECTOR_TAG}="${randId}"]`
+  }
+
+  async toPersistedRegions(driver) {
+    return locatorToPersistedRegions(this._selector, driver)
   }
 }